fix review comments
